+++ src/main/php/net/stubbles/webapp/UriPath.php
@@ -8,6 +8,7 @@
  * @package  net\stubbles\webapp
  */
 namespace net\stubbles\webapp;
+use net\stubbles\input\ValueReader;
 /**
  * Called path.
  *
@@ -75,6 +76,7 @@ class UriPath
      * @param   string  $name
      * @param   bool    $default
      * @return  string
+     * @deprecated  use readArgument() instead, will be removed with 4.0.0
      */
     public function getArgument($name, $default = null)
     {
@@ -85,6 +87,23 @@ class UriPath
         return $default;
     }
 
+
+    /**
+     * returns argument with given name or default if not set
+     *
+     * @param   string  $name
+     * @param   bool    $default
+     * @return  ValueReader
+     * @since   3.3.0
+     */
+    public function readArgument($name, $default = null)
+    {
+        if (isset($this->arguments[$name])) {
+            return ValueReader::forValue($this->arguments[$name]);
+        }
+
+        return ValueReader::forValue($default);
+    }
     /**
      * returns remaining path that was not matched by original path
      *